Fix MSVC bot input distribution types
Task-ID: MC-AI-0200

+++ src/client/BotClient.cpp
@@ -6,8 +6,8 @@ namespace client {
 
 shared::Direction BotClient::input() {
     static std::default_random_engine generator{ std::random_device{}() };
-    std::uniform_int_distribution<uint8_t> change_distribution{ 0U, 49U };
-    std::uniform_int_distribution<int8_t> direction_distribution{ -1, 1 };
+    std::uniform_int_distribution<int> change_distribution{ 0, 49 };
+    std::uniform_int_distribution<int> direction_distribution{ -1, 1 };
     if (change_distribution(generator) == 0) {
         m_direction.x = static_cast<uint8_t>(direction_distribution(generator) * 127);
         m_direction.y = static_cast<uint8_t>(direction_distribution(generator) * 127);
